requirements.txt
python-telegram-bot
thefuzz[speedup]
bs4
requests